Problem: SpiClient can be leaked from Spi::connect
This code will close the connection and return the client.

```rust
Spi::connect(|client| Ok(Some(client)))
```

If one will attempt to use it, it'll panic.

Solution: make the connection close on SpiClient drop

Also, make it impossible to create `SpiClient` out of thing air
to ensure there's always a connection to close.